# [aiohttp] - use fetchmany api