New asyncpg 0.30.0 [fetchmany](https://magicstack.github.io/asyncpg/current/api/index.html#asyncpg.connection.Connection.fetchmany) do the same stuff that we did in the queries test before, but much faster rather than calling fetchval in the loop. When queries count is low the results are the same, but when numbers of queries grow we can spot the difference.

### After
```
 Queries: 20 for query
 wrk -H 'Host: tfb-server' -H 'Accept: application/json,text/html;q=0.9,application/xhtml+xml;q=0.9,application/xml;q=0.8,*/*;q=0.7' -H 'Connection: keep-alive' --latency -d 30 -c 32 --timeout 8 -t 6 "http://tfb-server:8080/queries/20"
---------------------------------------------------------
Running 30s test @ http://tfb-server:8080/queries/20
  6 threads and 32 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     2.81ms    4.52ms 122.69ms   89.94%
    Req/Sec     2.91k     1.79k    7.96k    65.83%
  Latency Distribution
     50%    1.21ms
     75%    3.17ms
     90%    7.36ms
     99%   18.76ms
  522369 requests in 30.06s, 389.82MB read
Requests/sec:  17379.58
Transfer/sec:     12.97MB
```

### Before
```
---------------------------------------------------------
Running 30s test @ http://tfb-server:8080/queries/20
  6 threads and 32 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.54ms    7.84ms 161.52ms   96.87%
    Req/Sec     1.44k   445.63     2.97k    66.83%
  Latency Distribution
     50%    3.21ms
     75%    4.44ms
     90%    6.08ms
     99%   45.63ms
  258285 requests in 30.02s, 192.75MB read
Requests/sec:   8603.63
Transfer/sec:      6.42MB
STARTTIME 1745830666
ENDTIME 1745830696
```