@Reskov 2025-05-10
@Dreamsorcerer no, I am personally not interested in pypy3, but I can add a test if you want. Secondly, we may try to cythonize entire app code.

Do you have any thoughts why update performance is much lower than fastapi app have. 

https://www.techempower.com/benchmarks/#section=test&runid=6e3ad2d2-4acf-49cd-902a-286531ff3a67&f=zik0vz-zik0zj-zik0zj-zih7un-zik0zj-zik0zj-zik0zj-zik0zj-zik0zj-zik0zj-zik0zj-zik0zj-zik0zj-zik0zj-zik0zj-1ekf&test=query

<img width="1312" alt="image" src="https://github.com/user-attachments/assets/757ee1d0-2611-4d55-8a62-70f378c73361" />