@Dreamsorcerer 2025-05-10
> @Dreamsorcerer no, I am personally not interested in pypy3

Looks like FastAPI's fastest runs are on pypy3, so probably worth adding in. I think it's just a case of changing the docker image in the first line?

> Secondly, we may try to cythonize entire app code.

I think that's a step too far and no longer comparing frameworks. At some point, it'd be good to compile everything under mypyc though, which could easily be adopted to all frameworks (it's probably not quite mature enough yet to work on a general application though).