@Dreamsorcerer 2025-06-06
Fastapi still seems to get better performance on the updates. I'm wondering if one or both of the sorted() calls actually speeds up the DB query?

Pypy also seems to be slower for aiohttp for some reason..