@Reskov 2025-06-06
interestingly, looks like top performance framework are updating all rows at once. 

https://github.com/TechEmpower/FrameworkBenchmarks/blob/master/frameworks/Rust/viz/src/db_pg.rs#L62

https://github.com/TechEmpower/FrameworkBenchmarks/blob/master/frameworks/Java/inverno/src/main/java/com/techempower/inverno/benchmark/internal/DbRepository.java#L45