@Dreamsorcerer 2025-06-06
> interestingly, looks like top performance framework are updating all rows at once.

I assumed that's what we were doing already, but maybe I'm not familiar enough with the DB functions.