@Reskov 2025-06-06
@Dreamsorcerer not exactly I've prepared a draft here https://github.com/TechEmpower/FrameworkBenchmarks/pull/9932